Remove wrapped function to promised.js
