+++ src/api.js
@@ -1,56 +1,15 @@
 'use strict';
-const https = require('https');
-const readline = require('readline');
 const fs = require('fs');
-const { spawn } = require('child_process');
-
-const rl = readline.createInterface({
-  input: process.stdin,
-  output: process.stdout,
-});
+const promised = require('./promisificated.js');
 
 const green = '\x1b[32m';
 const red = '\x1b[31m';
 
-
-//Промисификация функций
-const question = str => new Promise(resolve => rl.question(str, resolve));
-const getRequest = async url => new Promise((resolve, reject) => {
-  https.get(url, async res => {
-    const buffers = [];
-    for await (const chunk of res) buffers.push(chunk);
-    const data = JSON.parse(Buffer.concat(buffers).toString());
-    if (data.Response === 'Error') reject(data.Message);
-    resolve(data);
-  })
-  //Чтобы отловить событие ошибки, потом нужно починить
-    .on('error', reject);
-});
-
-const postRequest = (options, data) => new Promise((resolve, reject) => {
-  const req = https.request(options, async res => {
-    const buffers = [];
-    for await (const chunk of res) buffers.push(chunk);
-    const reqData = JSON.parse(Buffer.concat(buffers).toString());
-    resolve(reqData);
-  });
-  req.write(data);
-  req.on('error', reject);
-  req.end();
-});
-
-const promiseSpawn = (lang, path) => new Promise((resolve, reject) => {
-  const pyProcess = spawn(lang, [path]);
-  pyProcess.stdout.on('data', data => resolve(JSON.parse(data)));
-  pyProcess.stderr.on('data', err => reject(err));
-});
-
-//Запись ответа в файл
 const writeFile = async resultTxt => {
-  const selection = parseInt(await question('Print 1 to save results\n'));
-  if (selection === 1) {
+  const select = parseInt(await promised.question('Print 1 to save results\n'));
+  if (select === 1) {
     const fileName = ('Write the name of txt file to save your results\n');
-    const txtName = await question(fileName);
+    const txtName = await promised.question(fileName);
     fs.writeFileSync(`${txtName}.txt`, resultTxt.join('\n'), 'utf8');
     return txtName;
 
@@ -60,21 +19,17 @@ const writeFile = async resultTxt => {
 
 };
 
-//Враппер для обработки ошибок
-const errorWrapper = handleError => func => (...args) =>
-  func(...args).catch(handleError);
-
 const handleError = e => {
   console.log(`Something gone wrong, error:\n${e}`);
   process.exit();
 };
 
-const errorHandlerWrapped = errorWrapper(handleError);
+const errorHandlerWrapped = promised.errorWrapper(handleError);
 
 //Обернутая функция request
-const safeGet = errorHandlerWrapped(getRequest);
-const safeSpawn = errorHandlerWrapped(promiseSpawn);
-const safePost = errorHandlerWrapped(postRequest);
+const safeGet = errorHandlerWrapped(promised.getRequest);
+const safeSpawn = errorHandlerWrapped(promised.promiseSpawn);
+const safePost = errorHandlerWrapped(promised.postRequest);
 
 class Crypto {
   constructor(key) {
@@ -83,8 +38,8 @@ class Crypto {
   }
 
   async currencyToCrypto() {
-    const currency = await question('Type currency you want to convert\n');
-    const query = this.defaultUrl + `/price?fsym=BTC&tsyms=${currency}`;
+    const curr = await promised.question('Type currency you want to convert\n');
+    const query = this.defaultUrl + `/price?fsym=BTC&tsyms=${curr}`;
     const result = await safeGet(query);
     const resultText = [];
     if (result) {
@@ -95,7 +50,6 @@ class Crypto {
       console.log(`${resultText.join('\n')}\n`);
       await writeFile(resultText);
     }
-    rl.close();
     return result;
   }
 
@@ -110,14 +64,13 @@ class Crypto {
     });
     console.log(`${resultText.join('\n')}\n`);
     await writeFile(resultText);
-    rl.close();
     return result;
   }
 
 
   async currencyPriceVolume() {
     const currText = 'Type curr you want to get 24h volume of/res curr\n';
-    const [curr, volumeCurr] = (await question(currText)).split(',');
+    const [curr, volumeCurr] = (await promised.question(currText)).split(',');
     const url = `/v2/histoday?fsym=${curr}&tsym=${volumeCurr}&limit=1`;
     const query = this.defaultUrl + url;
     const result = await safeGet(query);
@@ -139,7 +92,6 @@ class Crypto {
       await writeFile(resultText);
 
     }
-    rl.close();
     return result.Data;
   }
 
@@ -182,10 +134,7 @@ class Wallet {
 }
 
 module.exports = {
-  question,
   Crypto,
   Wallet,
-  getRequest,
   writeFile,
-  errorWrapper,
 };
